backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<4.4
